+++ EnergyOptimizer.Service/DTOs/ConsumptionTrendDto.cs
@@ -3,7 +3,7 @@
     public class ConsumptionTrendDto
     {
         public DateTime Timestamp { get; set; }
-        public double TotalConsumption { get; set; }
+        public decimal TotalConsumption { get; set; }
         public int ActiveDevices { get; set; }
     }
 }
