+++ EnergyOptimizer.Service/DTOs/HourlyConsumptionDto.cs
@@ -4,8 +4,8 @@
     {
         public int Hour { get; set; }
         public string TimeLabel { get; set; } = string.Empty;
-        public double TotalConsumption { get; set; }
+        public decimal TotalConsumption { get; set; }
         public int ReadingsCount { get; set; }
-        public double AverageConsumption { get; set; }
+        public decimal AverageConsumption { get; set; }
     }
 }
