+++ EnergyOptimizer.Service/DTOs/LiveReadingDto.cs
@@ -7,8 +7,8 @@
         public string DeviceType { get; set; } = string.Empty;
         public string ZoneName { get; set; } = string.Empty;
         public DateTime Timestamp { get; set; }
-        public double PowerConsumptionKW { get; set; }
-        public double Voltage { get; set; }
+        public decimal PowerConsumptionKW { get; set; }
+        public decimal Voltage { get; set; }
         public double Current { get; set; }
         public double Temperature { get; set; }
         public bool IsActive { get; set; }
