+++ EnergyOptimizer.Service/DTOs/UpdateDeviceDto.cs
@@ -7,7 +7,7 @@ namespace EnergyOptimizer.API.DTOs
         public string? Name { get; set; }
         public int? ZoneId { get; set; }
         public DeviceType? Type { get; set; }
-        public double? RatedPowerKW { get; set; }
+        public decimal? RatedPowerKW { get; set; }
         public bool? IsActive { get; set; }
     }
 }
